feat: enhance VariantStyleMixin with comprehensive variant functionality
- Remove duplicate StyleVariantMixin from variant_util.dart
- Update ImageStyler to use enhanced VariantStyleMixin for consistency
- Add 18 new variant methods to VariantStyleMixin:
  * Widget state variants: onError, onScrolledUnder, onDragged, onEnabled
  * Orientation variants: onPortrait, onLandscape
  * Device breakpoint variants: onMobile, onTablet, onDesktop
  * Text direction variants: onLtr, onRtl
  * Platform variants: onIos, onAndroid, onMacos, onWindows, onLinux, onFuchsia, onWeb
- Add comprehensive test coverage for all new variant methods
- Consolidate variant functionality into single source of truth
