+++ packages/mix/lib/src/specs/image/image_style.dart
@@ -8,7 +8,7 @@ import '../../core/style.dart';
 import '../../core/style_spec.dart';
 import '../../modifiers/modifier_config.dart';
 import '../../style/mixins/modifier_style_mixin.dart';
-import '../../variants/variant_util.dart';
+import '../../style/mixins/variant_style_mixin.dart';
 import 'image_spec.dart';
 import 'image_mutable_style.dart';
 import 'image_widget.dart';
@@ -19,7 +19,7 @@ class ImageStyler extends Style<ImageSpec>
     with
         Diagnosticable,
         ModifierStyleMixin<ImageStyler, ImageSpec>,
-        StyleVariantMixin<ImageStyler, ImageSpec> {
+        VariantStyleMixin<ImageStyler, ImageSpec> {
   final Prop<ImageProvider<Object>>? $image;
   final Prop<double>? $width;
   final Prop<double>? $height;
