+++ packages/mix/lib/src/style/mixins/variant_style_mixin.dart
@@ -74,8 +74,104 @@ mixin VariantStyleMixin<T extends Style<S>, S extends Spec<S>> on Style<S> {
     return variant(ContextVariant.widgetState(WidgetState.selected), style);
   }
 
+  /// Creates a variant for error state
+  T onError(T style) {
+    return variant(ContextVariant.widgetState(WidgetState.error), style);
+  }
+
+  /// Creates a variant for scrolled under state
+  T onScrolledUnder(T style) {
+    return variant(
+      ContextVariant.widgetState(WidgetState.scrolledUnder),
+      style,
+    );
+  }
+
+  /// Creates a variant for dragged state
+  T onDragged(T style) {
+    return variant(ContextVariant.widgetState(WidgetState.dragged), style);
+  }
+
+  /// Creates a variant for enabled state (opposite of disabled)
+  T onEnabled(T style) {
+    return variant(
+      ContextVariant.not(ContextVariant.widgetState(WidgetState.disabled)),
+      style,
+    );
+  }
+
   /// Creates a variant based on breakpoint
   T onBreakpoint(Breakpoint breakpoint, T style) {
     return variant(ContextVariant.breakpoint(breakpoint), style);
+  }
+
+  /// Creates a variant for portrait orientation
+  T onPortrait(T style) {
+    return variant(ContextVariant.orientation(Orientation.portrait), style);
+  }
+
+  /// Creates a variant for landscape orientation
+  T onLandscape(T style) {
+    return variant(ContextVariant.orientation(Orientation.landscape), style);
+  }
+
+  /// Creates a variant for mobile breakpoint
+  T onMobile(T style) {
+    return variant(ContextVariant.mobile(), style);
+  }
+
+  /// Creates a variant for tablet breakpoint
+  T onTablet(T style) {
+    return variant(ContextVariant.tablet(), style);
+  }
+
+  /// Creates a variant for desktop breakpoint
+  T onDesktop(T style) {
+    return variant(ContextVariant.desktop(), style);
+  }
+
+  /// Creates a variant for left-to-right text direction
+  T onLtr(T style) {
+    return variant(ContextVariant.directionality(TextDirection.ltr), style);
+  }
+
+  /// Creates a variant for right-to-left text direction
+  T onRtl(T style) {
+    return variant(ContextVariant.directionality(TextDirection.rtl), style);
+  }
+
+  /// Creates a variant for iOS platform
+  T onIos(T style) {
+    return variant(ContextVariant.platform(TargetPlatform.iOS), style);
+  }
+
+  /// Creates a variant for Android platform
+  T onAndroid(T style) {
+    return variant(ContextVariant.platform(TargetPlatform.android), style);
+  }
+
+  /// Creates a variant for macOS platform
+  T onMacos(T style) {
+    return variant(ContextVariant.platform(TargetPlatform.macOS), style);
+  }
+
+  /// Creates a variant for Windows platform
+  T onWindows(T style) {
+    return variant(ContextVariant.platform(TargetPlatform.windows), style);
+  }
+
+  /// Creates a variant for Linux platform
+  T onLinux(T style) {
+    return variant(ContextVariant.platform(TargetPlatform.linux), style);
+  }
+
+  /// Creates a variant for Fuchsia platform
+  T onFuchsia(T style) {
+    return variant(ContextVariant.platform(TargetPlatform.fuchsia), style);
+  }
+
+  /// Creates a variant for web platform
+  T onWeb(T style) {
+    return variant(ContextVariant.web(), style);
   }
 }
